% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskspecific.R
\name{find_task_data}
\alias{find_task_data}
\title{Find task data files}
\usage{
find_task_data(root_path)
}
\arguments{
\item{root_path}{Path of the root folder to start search.}
}
\value{
Character vector of path names.
}
\description{
Helper function to find a list of IWRT file names given a specific root folder.
}
\examples{
find_task_data('../data')
find_task_data('path/to/data/folder')
}

