% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskspecific.R
\name{find_tracking_data}
\alias{find_tracking_data}
\title{Find tracking data files}
\usage{
find_tracking_data(root_path)
}
\arguments{
\item{root_path}{Path of the root folder to start search.}
}
\value{
Character vector of path names.
}
\description{
Helper function to find a list of IWRT file names given a specific root folder.
}
\details{
If the root folder is provided this will find each of the separate .csv
files regarding the eye tracking data that exist for each trial 
and for each participant.
}
\examples{
find_tracking_data('../data')
find_tracking_data('path/to/data/folder')
}

