% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taskspecific.R
\name{import_task_data}
\alias{import_task_data}
\title{Import IWRT Task Data into R}
\usage{
import_task_data(root_path, file_paths, datatable = TRUE)
}
\arguments{
\item{root_path}{Main path to folder where all the individual subject folders are contained.}

\item{file_paths}{A character vector that lists specific file names to import.
The function \code{\link{find_task_data}} may be used to find exact file names.}
}
\value{
Will return a data.table containing the task data. Set \code{datatable=FALSE} to return a data.frame instead.
}
\description{
Imports the MATLAB task data into R from a folder or vector of filenames.
}
\details{
This function should be used if you only want the task data in raw format.
Use the function \code{\link{iwrt_auto}} to do a complete import and data processing.
}
\examples{
# import all data found in the upper level folder 'data'
import_tasks_data('../data')

# import data from specific files, use the default data.frame instead of data.table
import_task_data(file_paths=c('file1.csv', 'file2.csv'), datatable=FALSE)
}

