% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskspecific.R
\name{import_tracking_data}
\alias{import_tracking_data}
\title{Import IWRT Tracking Data into R}
\usage{
import_tracking_data(root_path, file_paths, datatable = TRUE)
}
\arguments{
\item{root_path}{Main path to folder where all the individual subject folders are contained.}

\item{file_paths}{A character vector that lists specific file names to import. 
The function \code{\link{find_tracking_data}} may be used to find exact file names.}
}
\value{
Will return a data.table containing the tracking data. Set \code{datatable=FALSE} to return a data.frame instead.
}
\description{
Imports the MATLAB tracking data into R from a folder or vector of filenames.
}
\details{
This function should be used if you only want the tracking data in raw format.
Use the function \code{\link{iwrt_auto}} to do a complete import and data processing.
}
\examples{
# import all data found in the upper level folder 'data'
import_tracking_data('../data')

# import data from specific files, use the default data.frame instead of data.table
import_tracking_data(file_paths=c('file1.csv', 'file2.csv'), datatable=FALSE)
}

