% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskspecific.R
\name{iwrt_auto}
\alias{iwrt_auto}
\title{Automatic IWRT data import and processing}
\usage{
iwrt_auto(root_path)
}
\arguments{
\item{root_path}{Character string of main path to folder where all the 
individual subject folders are contained.}
}
\value{
List of data.tables
}
\description{
Will automatically import and processes data given root path
}
\details{
Must specify package \code{options()} if different from defaults
}
\examples{
# Only need to specify the root folder where data is contained
data_list <- iwrt_auto("../data")

# Extract specific datasets from the list
roi_data <- data_list$roi
all_data <- data_list$complete
}

