% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taskspecific.R
\name{iwrt_import_raw}
\alias{iwrt_import_raw}
\title{Import raw data}
\usage{
iwrt_import_raw(root_path)
}
\arguments{
\item{root_path}{Character string of path to the root folder where
data subfolders are contained}
}
\value{
Returns a data.table of merged task and tracking data
}
\description{
Imports raw data with no processing and all columns from task and tracking
}
\details{
This is to get both tracking and task data merged as one complete set without
doing any further processing of the data. Use \code{\link{iwrt_auto}} for
an automatic method for import and processing of ROI's and tracking columns.
}
\examples{
iwrt_import_raw("../data")
}

