% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskspecific.R
\name{iwrt_merge}
\alias{iwrt_merge}
\title{Merge task and tracking data}
\usage{
iwrt_merge(task, tracking)
}
\arguments{
\item{task}{task data. If not a data.table will be converted to one}

\item{tracking}{tracking data. If not a data.table will be converted to one}
}
\value{
data.table of merged tracking and task data
}
\description{
Merges two different data sets of eye gaze data and task related data
}
\examples{
task <- import_task_data(root_path)
tracking <- import_tracking_data(root_path)
merged_data <- iwrt(task, tracking)
}

