% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taskspecific.R
\name{iwrt_roi}
\alias{iwrt_roi}
\title{ROI classification}
\usage{
iwrt_roi(in_dat, xname, yname)
}
\arguments{
\item{in_dat}{data.table with the ROI column and their normalized coordinates}

\item{xname}{normalized coordinate for the x axis of the eye gaze position}

\item{yname}{normalized coordinate for the y axis of the eye gaze position}
}
\value{
Same data.table but we new columns added for each ROI
}
\description{
Classifies ROIs based on coordinates
}
\details{
Must use \code{options()} to set the column names of your ROIs. Defaults to
  left and right image locations. xname and yname are eyegaze positions.
}

