% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r.iwrt.R
\docType{package}
\name{r.iwrt}
\alias{r.iwrt}
\alias{r.iwrt-package}
\title{r.iwrt: A Package for Analysis of MATLAB IWRT Data}
\description{
The r.iwrt packages provides a set of automated tools and functions to analyze
the Infant Word Recognition Task data obtained from the MATLAB program.
}
\section{Task specific functions}{

\code{\link{iwrt_auto}} \cr
\code{\link{find_task_data}} \cr
\code{\link{find_tracking_data}} \cr
\code{\link{import_task_data}} \cr
\code{\link{import_tracking_data}} \cr
\code{\link{iwrt_import_raw}} \cr
}

\section{Tobii eye tracker functions}{

\code{\link{tracking_process}} \cr
}

