% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eyetrack.R
\name{tracking_process}
\alias{tracking_process}
\title{Process tracking data}
\usage{
tracking_process(in_dat, which_axis)
}
\arguments{
\item{in_dat}{data.table with necessary tracking columns. Will attempt to find
them from the package options in \code{options()}}

\item{which_axis}{character string indicating which axis to process. Can be one
  of \code{c("x","y","horz","vert")}}
}
\value{
Numeric vector of eyetracker coordinates for one of the specified axes
}
\description{
A series of operations to filter and processing eye tracking data
}
\details{
Must use \code{options} to set parameters for how to process the data. This
  function is used automatically by the main function \code{\link{iwrt_auto}}.
}
\examples{
tracking_process(my_raw_data, "x")
}

